Desktop/projects/PROJECT 3/files/code-review-agent/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
langgraph==0.1.19
langchain-groq==0.1.9
langchain-core==0.2.38
PyGithub==2.3.0
pydantic==2.7.4
